Fix stitch issue encountered in client code
